dont initialise in protocol
